Remove V1/V2 distinction, separate migration tests
